+++ tests/odoo-tests.sh
@@ -1,17 +1,17 @@
 ## Odoo-WooCommerce Sync Tests
-# Last update: 2025-03-02
+# Last update: 2025-05-22
 
 
 # Settings
 website="website.com"
-website_root_directory="/var/www/vhosts/$website/httpdocs"
+website_root_path="/var/www/vhosts/$website/httpdocs"
 odoo_conf="/etc/odoo.conf"
 database_name="database_name"
 odoo_addon_name="woocommerce_sync"
 
 
 # First Odoo run after fresh install
-# $website_root_directory/odoo/venv/bin/python3 $website_root_directory/odoo/odoo-bin --config=$odoo_conf --database $database_name --load=base,web --without-demo=all --update=all
+# $website_root_path/odoo/venv/bin/python3 $website_root_path/odoo/odoo-bin --config=$odoo_conf --database $database_name --load=base,web --without-demo=all --update=all
 
 
 
@@ -26,38 +26,38 @@ ps aux | grep odoo | grep -v grep | awk '{print $2}' | xargs -r kill -9
 
 
 # Start Odoo command line
-# $website_root_directory/odoo/venv/bin/python3 $website_root_directory/odoo/odoo-bin shell --config=$odoo_conf
+# $website_root_path/odoo/venv/bin/python3 $website_root_path/odoo/odoo-bin shell --config=$odoo_conf
 
 # Retrieve all field information for product templates
 # fields = self.env['product.template'].fields_get()
-# fields
+# print(fields)
 
 # Retrieve default values for product template fields
 # print(self.env['product.template'].default_get(self.env['product.template']._fields.keys()))
 
 # Retrieve required fields
 # fields = {field: data for field, data in fields.items() if data.get('required')}
-# fields.keys()
+# print(fields.keys())
 
 # Retrieve read-only fields
 # fields = {field: data for field, data in fields.items() if data.get('readonly')}
-# fields.keys()
+# print(fields.keys())
 
 
 # Update the WooCommerce Sync module in Odoo and stop Odoo after initialization
-$website_root_directory/odoo/venv/bin/python3 $website_root_directory/odoo/odoo-bin \
+$website_root_path/odoo/venv/bin/python3 $website_root_path/odoo/odoo-bin \
     --config=$odoo_conf \
     --database $database_name \
     --update $odoo_addon_name \
     --stop-after-init
 
 # Overwrite the Odoo log file with an empty content
-cat <<EOF > "/var/log/odoo.log"
+cat <<EOF > "/var/log/odoo/$website.log"
 EOF
 
 # Update WooCommerce Sync module and load additional dependencies
-$website_root_directory/odoo/venv/bin/python3 $website_root_directory/odoo/odoo-bin \
+$website_root_path/odoo/venv/bin/python3 $website_root_path/odoo/odoo-bin \
 	--config=$odoo_conf \
 	--database $database_name \
 	--update $odoo_addon_name \
-	--load=product
+	--load=product,web,queue_job